feat. ajustes finalizados para deploy

+++ src/services/deleteUserService.ts
@@ -4,6 +4,13 @@ import { deleteCredentialRepository, getAllCredentialsByIdRepository } from "../
 
 export async function deleteUserService(id: number) {
   const userCredentials = await getAllCredentialsByIdRepository(id)
+  console.log(userCredentials)
+  if(userCredentials === null || userCredentials.length === 0) {
+    throw {
+      type: "Unauthorized",
+      message: "Usuário não existe"
+    }
+  }
   const promises = userCredentials.map(item => {return deleteCredentialRepository(item.id)})
   await Promise.all(promises)
   const result = await repository.deleteUser(id)
